// src/app/admin/interfaces/eventos.interface.ts
// Generated by https://quicktype.io

import { AreasConocimiento } from "./areasConocimiento.interfaace";

export interface Eventos {
  id:          number;
  titulo:      string;
  descripcion: string;
  duracion:    string;
  fecha:       string;
  imgen:       string;
  estado:      'Sin Asignación' | 'Pendiente Asignación' | 'Listo';
  inicioHora:  string;
  areas:       AreasConocimiento[];
}